product-insight-reveal-main/backend/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
aiofiles>=23.2.0
urllib3>=2.0.0
fake-useragent>=1.4.0 
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.6 